@@ -17,7 +17,7 @@
     "start": "node demo",
     "pretest": "npm run clean && npm run build",
     "test": "run-p -r start-http-server 'run-tests --scripts-prepend-node-path'",
-    "run-tests": "wait-on http://localhost:9615 && jest",
+    "run-tests": "wait-on http://localhost:9615 && vitest run",
     "start-http-server": "http-server test/fixtures --port=9615 --silent",
     "lint": "eslint --ext=ts,js --ignore-path .gitignore .",
     "build": "./scripts/generate-exports.js && tsc",
@@ -25,36 +25,40 @@
     "prepare": "husky install"
   },
   "dependencies": {
-    "@types/pngjs": "^6.0.1",
-    "aws-sdk": "^2.1354.0",
+    "@aws-sdk/client-device-farm": "^3.354.0",
+    "@aws-sdk/credential-providers": "^3.354.0",
     "get-stream": "^6.0.1",
     "lodash": "^4.17.21",
     "p-retry": "^4.6.2",
     "pixelmatch": "^5.3.0",
-    "pngjs": "^6.0.0",
-    "webdriverio": "^7.25.2"
+    "pngjs": "^7.0.0",
+    "webdriverio": "^8.11.2"
   },
   "devDependencies": {
-    "@types/jest": "^29.2.0",
+    "@types/debug": "^4.1.8",
     "@types/lodash": "^4.14.186",
-    "@types/node": "^16.0.0",
+    "@types/node": "^20.3.1",
     "@types/pixelmatch": "^5.2.4",
+    "@types/pngjs": "^6.0.1",
     "@typescript-eslint/eslint-plugin": "^5.41.0",
     "@typescript-eslint/parser": "^5.41.0",
+    "@vitest/coverage-istanbul": "^0.32.2",
+    "@wdio/globals": "^8.11.2",
+    "aws-sdk-client-mock": "^2.1.1",
     "eslint": "^8.26.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-header": "^3.1.1",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-unicorn": "^44.0.2",
+    "eslint-plugin-unicorn": "^47.0.0",
     "http-server": "^14.1.1",
     "husky": "^8.0.1",
     "jest": "^29.2.2",
     "lint-staged": "^13.0.3",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.7.1",
-    "ts-jest": "^29.0.3",
-    "typescript": "^4.8.4",
-    "wait-on": "^6.0.1"
+    "typescript": "^5.1.3",
+    "vitest": "^0.32.2",
+    "wait-on": "^7.0.1"
   },
   "exports": {
     "./chrome-launcher": "./chrome-launcher.js",
