@@ -1,6 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-const BrowserStackBrowserCreator = require('../../src/browsers/browserstack').default;
+import { beforeEach, describe, test, expect, vi } from 'vitest';
+
+import BrowserStackBrowserCreator from '../../src/browsers/browserstack';
 
 const browserName = 'Chrome';
 const browserOptions = {
@@ -11,7 +13,7 @@ const browserOptions = {
   nSecSleepBeforeRetry: 0.001,
 };
 
-BrowserStackBrowserCreator.prototype.setupBrowser = jest.fn().mockImplementation(options => {
+(BrowserStackBrowserCreator as any).prototype.setupBrowser = vi.fn().mockImplementation(options => {
   const errorMessage = options.makeQueueFullErrorForTesting
     ? 'All parallel tests are currently in use, including the queued tests. Please wait to finish or upgrade your plan to add more sessions.'
     : 'Some other error that we should accept without retrying';
@@ -21,23 +23,23 @@ BrowserStackBrowserCreator.prototype.setupBrowser = jest.fn().mockImplementation
 
 describe('BrowserStack browserCreator ', () => {
   beforeEach(() => {
-    jest.clearAllMocks();
+    vi.clearAllMocks();
   });
 
   test('should try to reconnect in case of a full BrowserStack queue', async () => {
     const browserCreator = new BrowserStackBrowserCreator(browserName, browserOptions);
-    expect(browserCreator.setupBrowser.mock.calls.length).toBe(0);
-    await expect(browserCreator.getBrowser({ makeQueueFullErrorForTesting: true })).rejects.toThrow(
+    expect((browserCreator as any).setupBrowser.mock.calls.length).toBe(0);
+    await expect(browserCreator.getBrowser({ makeQueueFullErrorForTesting: true } as any)).rejects.toThrow(
       'All parallel tests are currently in use'
     );
-    expect(browserCreator.setupBrowser.mock.calls.length).toBe(11);
+    expect((browserCreator as any).setupBrowser.mock.calls.length).toBe(11);
   });
 
   test('should not try to reconnect if any other error than a full BrowserStack queue gets thrown', async () => {
     const browserCreator = new BrowserStackBrowserCreator(browserName, browserOptions);
-    expect(browserCreator.setupBrowser.mock.calls.length).toBe(0);
+    expect((browserCreator as any).setupBrowser.mock.calls.length).toBe(0);
     await expect(browserCreator.getBrowser({})).rejects.toThrow();
-    expect(browserCreator.setupBrowser.mock.calls.length).toBe(1);
+    expect((browserCreator as any).setupBrowser.mock.calls.length).toBe(1);
   });
 
   test('should use BrowserStack hub url', async () => {
