@@ -1,9 +1,12 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-const fs = require('fs');
-const path = require('path');
-const mergeImages = require('../src/image-utils/merge').default;
-const { parsePng } = require('../src/image-utils');
+import fs from 'node:fs';
+import path from 'node:path';
+
+import { describe, test, expect } from 'vitest';
+
+import mergeImages from '../src/image-utils/merge';
+import { parsePng } from '../src/image-utils';
 
 describe('mergeImages', () => {
   test('should not merge images beyond their dimensions', async () => {
