@@ -1,5 +1,3 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-module.exports = {
-  chromeDriverPort: '9515',
-};
+export const chromeDriverPort = '9515';
