@@ -1,5 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-const { shutdownWebdriver } = require('../../dist/chrome-launcher');
+import { shutdownWebdriver } from '../../dist/chrome-launcher';
 
-module.exports = shutdownWebdriver;
+export default shutdownWebdriver;
