@@ -2,13 +2,15 @@
 // SPDX-License-Identifier: Apache-2.0
 import { URL } from 'url';
 import pRetry, { AbortError } from 'p-retry';
-import DeviceFarm from 'aws-sdk/clients/devicefarm';
-import { CredentialsOptions } from 'aws-sdk/lib/credentials';
-import { AWSError } from 'aws-sdk/lib/error';
+
+import { AwsAuthInputConfig } from '@aws-sdk/middleware-signing';
+import { DeviceFarmClient, CreateTestGridUrlCommand, DeviceFarmServiceException } from '@aws-sdk/client-device-farm';
+
+import type { Capabilities } from '@wdio/types';
+
 import BrowserCreator from './browser-creator';
 import defaultCapabilities, { getCapability } from './capabilities';
 import { FatalError } from '../exceptions';
-import type { Capabilities } from '@wdio/types';
 
 // This uses undocumented options which prevents proper typing
 const capabilities: Record<string, Capabilities.DesiredCapabilities> = {
@@ -22,26 +24,25 @@ export interface DevicefarmOptions {
   retryCount?: number;
   projectArn: string;
   expiresInSeconds?: number;
-  credentials: CredentialsOptions;
+  credentials: AwsAuthInputConfig['credentials'];
 }
 
 export default class DevicefarmBrowserCreator extends BrowserCreator {
   protected async __getBrowserUrl(): Promise<URL> {
     const options = this.options as DevicefarmOptions;
-    const client = new DeviceFarm({ credentials: options.credentials, region: 'us-west-2' });
+    const client = new DeviceFarmClient({ credentials: options.credentials, region: 'us-west-2' });
 
     const response = await pRetry(
       async () => {
         try {
-          const response = await client
-            .createTestGridUrl({
+          return await client.send(
+            new CreateTestGridUrlCommand({
               projectArn: options.projectArn,
               expiresInSeconds: options.expiresInSeconds || 300,
             })
-            .promise();
-          return response;
+          );
         } catch (error) {
-          if (isAWSError(error) && error.code === 'ThrottlingException') {
+          if (error instanceof DeviceFarmServiceException && error.$retryable?.throttling) {
             // Retries with backoff when throttled
             throw error;
           }
@@ -54,14 +55,11 @@ export default class DevicefarmBrowserCreator extends BrowserCreator {
     if (!response.url) {
       throw new FatalError('Invalid response from devicefarm: ' + JSON.stringify(response));
     }
+
     return new URL(response.url);
   }
 
   protected __getCapabilities() {
     return getCapability(this.browserName, capabilities);
   }
-}
-
-function isAWSError(obj: unknown): obj is AWSError {
-  return obj instanceof Error && typeof (obj as { code?: string })['code'] === 'string';
 }
